fmt: Choose only gcc to execute a specific portion of code
The specific portion of the code needs to be executed by a gcc
compiler, for that reason is used the __GNUC__ macro. It appears,
that the LLVM compiler, which uses the __llvm__ macro, uses as well
the __GNUC__ macro (possible for compatibility???).

This change check if the __llvm__ is check as well, and if it is
the code is not executed.

Issue: https://github.com/svagionitis/cross-platform-code/issues/1